FormGenerator working, Registration nearly finished
FormGenerator is now able to create and sanitize Email-Input, Text-Input
and Password-Input including trim and HTMLSpecialchars.

The Registration Module is nearly finished, an implementation of
Submodel Accounts is necessary to finally create an account.

Modules can now block or unblock the content of a page. The blocking of
all other modules has to be implemented later.