Added Submodel Accounts and \Query\InsertInto
It is now possible to insert values into the database using
$model->insertInto().
The class \Query\SQLFunction provides a container for SQL functions to
use in \Query\Select or \Query\InsertInto.
The Submodel Accounts was added, beeing now able to check for existing
email addresses and occupied accounts.names and to create Accounts.
The Registration-Localmodule is now able to make use of the
Accounts-Submodel and creates an Account if the form is accepted. It
lacks, however, proper success/fail messages.