New Page, custom navigation
It is now possible to create a new page using the Tableedit-Module.
Additionally, the \Navigation\Container class was extended to allow
Custom, non-db-based navigation items. Localmodules got an additional
hook (navigationHook) where the navigation should be changed. It is
triggered the same way the hooks execute() and output() are triggered.