Added JS range indicator
Added a output-field for ranged input which gets refreshed with JS every
time the range value changes.
Additionally, the display of NULL-values in TableGenerator was adjusted
to actually write "NULL" inside the table - marked with <i
class="isnull"> in order to change the appearance with CSS. NULL gets
only written (in the module Tableedit) if the validator "nullifempty" is
set to true.