Character Creation
Further preparations for character creation: Implemented a simple form
renderer on client side that is able to submit the form back to the
server.

CharacterController uses same route for POST and GET and checks on its
own which one it uses - sends the form data on GET and an error 422 at
POST.

Furthermore, fixed the htaccess fiel to redirect via 308 to keep request
method the same.

ToDo: Implement FormValidation