Character Selection
The ability to select (and switch) the current character has been
implemented. The CharacterController checks if the user is the owner of
the chosen character - as well as if it is already in use by someone
else (if, later, character sharing has been implemented?)

The JS-Client supports the selection of a character.