SceneController: Loading default Scene
Added a Scene-Controller that is capable of returning the default scene
(for now: ID = 1).

Added 2 Tables: scenes and character_scenes. Scenes is for raw scenes as
defined by the server, character_scenes is after running modules (in
order to add content).

ToDo: Scene-Display in client + "actions" that are attached to the
scenes.